requirements.txt
streamlit
scikit-learn
tensorflow
numpy
pandas
matplotlib
seaborn
Pillow
plotly